Implement communication between PyQt event loop other processes
The event loop now periodically checks for any messages from
other process by checking the queue. It can then decide which
action to take and accordingly change the GUI. This can also
be used to send periodic messages but its main intent is to
check for messages. Using this it will be possible for the
Controller process to instruct View.