Extend post_template functionality
Set up interaction with the access rights buttons in the post
template. Clicking the particular access right button inside
a post_template changes the background color indicating that
the aceess right is active or not. Further functionality like
storing the status value of the access right and updating it
in database (via a call to controller) can also be added.

At present the widgets inside the post_template is fixed and
perform a simple task of changing background color, but in
further commits option to add new access right from a list
of available one can be added. Also graphics for those will
also be changed.