Make performance improvements
ModelHandler and ControllerHandler both used a busy loop which checked
for any new messages. Due to this there was huge wastae of CPU cycles
leading to high CPU utilization even if all it did was to make a simple
check for messages. Instead of this now both the handler waits for new
message on their respective queues and as soon as they get the message
it is identified and executed. This also reduces latency becuase as soon
as message is recieved it is catched.

A major problem though is that ModelHandler and ControllerHandler gets
halted and cannot do anything until they recieve a message. This might
become an issue when controller/model needs to do some task independent
of the other as they will be waiting for message. A solution can be to
send a special message from the ViewHandler which will tell
ModelHandler and ControllerHandler to do their routine check and after
that they can again wait for message. This special message can be send
periodically so as to keep everything updated. Other than this approach
we can use asynchronous programming but we do not have experience with it
as of now so we will use the prior. But if we see a possible performance
issue or other issues asynchronous will be a possible approach.