Make size for send messages dynamic
If the text size of message is small enough then the wordwrapping is
turned off so that the message is displayed on the rightmost side.
The messages size changes on resizing of window, so if on resizing the
text size becomes smaller or larger then wordwrapping is turned off or
on accordingly.

Since all the message widgets are checked on resizing of window it might
become a very computational heavy task and might take considerable time
to execute hence affecting the efficiency and user experience. One
workaround can be to do this resizing on only those messages that are
currently visible in the central window.
Another problem is that bigger messages don't resize on window resize.