Create classes for implementing class diagram
UserFactory class will be responsible for creating the instance of a
user based on the rquirements (access rights). It will create an
instance of BasicOperationHandler and according to the need add
functionality by providing a delegatee which provides additional
responsiblities.
It sort of uses Chain of Responsiblity pattern. If the object has the
required methods then it calls them otherwise it delegates the request
to its delegatee which then further does the same procedure until a
method is found or it has reached the end of chain and hence raise an
exception.
To provide flexiblity in adding new features Strategy pattern. Different
strategies can be used and changed dynamically based on changing
requirements.
Refer to class diagram for more information.