Implement send and receive from ui
Updated client.py to include a class that manages continues connection
to the server so as to send and receive messages. Used queues so as to
easily and safely transfer data between the ui and the backend. The
queue helps to easily check for new messages rather than using some
trivial strategy like flags.
The class provides a nice abstraction of connection and data transfer
between client and server.

Updated ui file to periodically check for new messages by looking at
the provided queue and accordingly make updates in GUI.