Feature to show chat/person details
Clicking on the particular chat heading opens up a window showing
details. A call is made to method of root window which then switches the
layout to show details based on provided chat_id. On clicking close
button on the chat details screen, again a call is made to root window
which then switches back to normal chat.
Passed root object so that child can call root object's methods.